use helper functions only when needed
